[chore #161637615] Put my in different files for each category
