+++ app/api/v1/views.py
@@ -31,7 +31,7 @@ def token_auth(func):
             )
         current_user = None
         try:
-            token_data = jwt.decode(token, Config.secret_key)
+            token_data = jwt.decode(token, Config.secret_key, algorithms=["HS256"])
         except:
             return make_response(
                 jsonify({"message": "invalid token"}), 401
